Add WithT method on TestDB. Update folder_test to use published package.

+++ lambda/restore/handler/file_test.go
@@ -96,7 +96,7 @@ func TestNameParts_Limit(t *testing.T) {
 
 func TestRestoreName(t *testing.T) {
 	db := store.OpenDB(t)
-	defer db.Close()
+	t.Cleanup(func() { db.Close() })
 	orgId := 2
 	for name, d := range map[string]struct {
 		id             int64
@@ -123,11 +123,13 @@ func TestRestoreName(t *testing.T) {
 			NodeId: d.nodeId,
 			Name:   d.name,
 		}
-		orginalName, err := GetOriginalName(d.name, d.nodeId)
+		originalName, err := GetOriginalName(d.name, d.nodeId)
 		if err != nil {
 			assert.FailNow(t, "test case does not use correct deleted file name format", err)
 		}
 		t.Run(name, func(t *testing.T) {
+			subDB := db.WithT(t)
+			t.Cleanup(func() { subDB.Truncate(orgId, "packages") })
 			var restoredName *RestoredName
 			err := messageHandler.Store.SQLFactory.ExecStoreTx(ctx, orgId, func(store store.SQLStore) (restoreNameError error) {
 				restoredName, restoreNameError = messageHandler.restoreName(ctx, restoreInfo, store)
@@ -136,19 +138,18 @@ func TestRestoreName(t *testing.T) {
 			if assert.NoError(t, err) {
 				query := fmt.Sprintf(`SELECT name from "%d".packages where id = $1`, orgId)
 				var actualName string
-				err = db.QueryRow(query, restoreInfo.Id).Scan(&actualName)
+				err = subDB.QueryRow(query, restoreInfo.Id).Scan(&actualName)
 				if assert.NoError(t, err) {
 					assert.Equal(t, d.expectedResult, actualName)
 					assert.Equal(t, d.expectedResult, restoredName.Value)
-					if actualName == orginalName {
+					if actualName == originalName {
 						assert.Empty(t, restoredName.OriginalName)
 					} else {
-						assert.Equal(t, orginalName, restoredName.OriginalName)
+						assert.Equal(t, originalName, restoredName.OriginalName)
 					}
 				}
 			}
 		})
-		db.Truncate(orgId, "packages")
 
 	}
 }
@@ -210,7 +211,11 @@ func TestMessageHandler_handleFilePackage(t *testing.T) {
 	ctx := context.Background()
 
 	db := store.OpenDB(t)
+	t.Cleanup(func() {
+		db.Close()
+	})
 	orgId := 2
+	datasetId := 1
 
 	awsConfig := store.GetTestAWSConfig(t)
 
@@ -225,73 +230,100 @@ func TestMessageHandler_handleFilePackage(t *testing.T) {
 		options.BaseEndpoint = aws.String(store.GetTestDynamoDBURL())
 	})
 
-	t.Cleanup(func() {
-		db.Truncate(orgId, "packages")
-		db.Truncate(orgId, "package_storage")
-		db.Truncate(orgId, "dataset_storage")
-		db.TruncatePennsieve("organization_storage")
-		db.Close()
-	})
-
-	datasetId := 1
 	bucketName := "test-bucket-handle-file"
 
-	testPackage := NewTestSourcePackage(1, datasetId, 1, func(testPackage *store.TestPackage) {
-		testPackage.Restoring()
-	}).WithSources(1, bucketName, func(testFile *store.TestFile) {
-		testFile.WithPublished(false)
-	}).Insert(ctx, db, orgId)
-	putObjectInputs := testPackage.PutObjectInputs()
+	for scenario, tt := range map[string]struct {
+		sourcePackage *TestSourcePackage
+	}{
+		"unpublished file package": {
+			sourcePackage: NewTestSourcePackage(1, datasetId, 1, func(testPackage *store.TestPackage) {
+				testPackage.Restoring()
+			}).WithSources(1, bucketName, func(testFile *store.TestFile) {
+				testFile.WithPublished(false)
+			}),
+		},
+		"published file package": {
+			sourcePackage: NewTestSourcePackage(2, datasetId, 1, func(testPackage *store.TestPackage) {
+				testPackage.Restoring()
+			}).WithSources(3, bucketName, func(testFile *store.TestFile) {
+				testFile.WithPublished(true)
+			}),
+		},
+	} {
 
-	s3Fixture := store.NewS3Fixture(t, s3Client, &s3.CreateBucketInput{Bucket: aws.String(bucketName)}).
-		WithBucketVersioning(bucketName).
-		WithObjects(putObjectInputs...)
-	defer s3Fixture.Teardown()
+		t.Run(scenario, func(t *testing.T) {
+			subDB := db.WithT(t)
+			t.Cleanup(func() {
+				subDB.Truncate(orgId, "packages")
+				subDB.Truncate(orgId, "package_storage")
+				subDB.Truncate(orgId, "dataset_storage")
+				subDB.TruncatePennsieve("organization_storage")
+			})
+			tt.sourcePackage.Insert(ctx, subDB, orgId)
+			putObjectInputs := tt.sourcePackage.PutObjectInputs()
 
-	keyToDeleteVersionId := testPackage.DeleteFiles(ctx, t, s3Client)
+			s3Fixture := store.NewS3Fixture(t, s3Client, &s3.CreateBucketInput{Bucket: aws.String(bucketName)}).
+				WithBucketVersioning(bucketName).
+				WithObjects(putObjectInputs...)
+			t.Cleanup(func() {
+				s3Fixture.Teardown()
+			})
 
-	putItemInputs := testPackage.PutItemInputs(t, deleteRecordTableName, keyToDeleteVersionId)
+			keyToDeleteVersionId := tt.sourcePackage.DeleteFiles(ctx, t, s3Client)
 
-	createTableInput := store.TestCreateDeleteRecordTableInput(deleteRecordTableName)
-	dyFixture := store.NewDynamoDBFixture(t, dyClient, &createTableInput).WithItems(putItemInputs...)
-	defer dyFixture.Teardown()
+			putItemInputs := tt.sourcePackage.PutItemInputs(t, deleteRecordTableName, keyToDeleteVersionId)
 
-	sqlFactory := store.NewPostgresStoreFactory(db.DB)
-	dyStore := store.NewDynamoDBStore(dyClient, deleteRecordTableName)
-	objectStore := store.NewS3Store(s3Client)
-	handler := NewMessageHandler(events.SQSMessage{MessageId: uuid.NewString(), Body: "{}"}, NewBaseStore(sqlFactory, dyStore, objectStore, nil))
-	restoreInfo := models.RestorePackageInfo{
-		Id:     testPackage.Package.Id,
-		NodeId: testPackage.Package.NodeId,
-		Name:   testPackage.Package.Name,
-		Type:   testPackage.Package.PackageType,
+			createTableInput := store.TestCreateDeleteRecordTableInput(deleteRecordTableName)
+			dyFixture := store.NewDynamoDBFixture(t, dyClient, &createTableInput).WithItems(putItemInputs...)
+			t.Cleanup(func() {
+				dyFixture.Teardown()
+			})
+
+			sqlFactory := store.NewPostgresStoreFactory(subDB.DB)
+			dyStore := store.NewDynamoDBStore(dyClient, deleteRecordTableName)
+			objectStore := store.NewS3Store(s3Client)
+			handler := NewMessageHandler(events.SQSMessage{MessageId: uuid.NewString(), Body: "{}"}, NewBaseStore(sqlFactory, dyStore, objectStore, nil))
+			restoreInfo := models.RestorePackageInfo{
+				Id:     tt.sourcePackage.Package.Id,
+				NodeId: tt.sourcePackage.Package.NodeId,
+				Name:   tt.sourcePackage.Package.Name,
+				Type:   tt.sourcePackage.Package.PackageType,
+			}
+			changelogEvents, err := handler.handleFilePackage(ctx, orgId, int64(datasetId), restoreInfo)
+			require.NoError(t, err)
+			assert.Len(t, changelogEvents, 1)
+			changelogEvent := changelogEvents[0]
+			assert.Equal(t, tt.sourcePackage.Package.Id, changelogEvent.Id)
+			assert.Equal(t, tt.sourcePackage.Package.NodeId, changelogEvent.NodeId)
+			assertRestoredName(t, tt.sourcePackage.Package.NodeId, tt.sourcePackage.Package.Name, changelogEvent.Name)
+			assert.Empty(t, changelogEvent.OriginalName)
+			assert.Nil(t, changelogEvent.Parent)
+
+			// no storage tables values have been set up, so the storage now should just be the size of the package
+			packageSize := tt.sourcePackage.Size()
+			actualPackageStorage := subDB.GetPackageStorage(orgId, int(tt.sourcePackage.Package.Id))
+			assert.Equal(t, actualPackageStorage, packageSize)
+			assert.Equal(t, subDB.GetDatasetStorage(orgId, datasetId), packageSize)
+			assert.Equal(t, subDB.GetOrganizationStorage(orgId), packageSize)
+
+			v := subDB.Queries(orgId)
+			actual, err := v.GetPackageByNodeId(ctx, tt.sourcePackage.Package.NodeId)
+			require.NoError(t, err)
+			assertRestoredPackage(t, FileRestoredState, tt.sourcePackage.Package.AsPackage(), actual)
+
+			listOut, err := s3Fixture.Client.ListObjectVersions(ctx, &s3.ListObjectVersionsInput{Bucket: aws.String(bucketName)})
+			require.NoError(t, err)
+			// No more delete markers and the number of versions is the same.
+			assert.Empty(t, listOut.DeleteMarkers)
+			assert.Len(t, listOut.Versions, len(putObjectInputs))
+
+			scanOut, err := dyFixture.Client.Scan(ctx, &dynamodb.ScanInput{TableName: aws.String(deleteRecordTableName)})
+			require.NoError(t, err)
+			// All delete records have been removed
+			assert.Zero(t, scanOut.ScannedCount)
+			assert.Empty(t, scanOut.Items)
+		})
 	}
-	changelogEvents, err := handler.handleFilePackage(ctx, orgId, int64(datasetId), restoreInfo)
-	require.NoError(t, err)
-	assert.Len(t, changelogEvents, 1)
-	changelogEvent := changelogEvents[0]
-	assert.Equal(t, testPackage.Package.Id, changelogEvent.Id)
-	assert.Equal(t, testPackage.Package.NodeId, changelogEvent.NodeId)
-	assertRestoredName(t, testPackage.Package.NodeId, testPackage.Package.Name, changelogEvent.Name)
-	assert.Empty(t, changelogEvent.OriginalName)
-	assert.Nil(t, changelogEvent.Parent)
-
-	v := db.Queries(orgId)
-	actual, err := v.GetPackageByNodeId(ctx, testPackage.Package.NodeId)
-	require.NoError(t, err)
-	assertRestoredPackage(t, FileRestoredState, testPackage.Package.AsPackage(), actual)
-
-	listOut, err := s3Fixture.Client.ListObjectVersions(ctx, &s3.ListObjectVersionsInput{Bucket: aws.String(bucketName)})
-	require.NoError(t, err)
-	// No more delete markers and the number of versions is the same.
-	assert.Empty(t, listOut.DeleteMarkers)
-	assert.Len(t, listOut.Versions, len(putObjectInputs))
-
-	scanOut, err := dyFixture.Client.Scan(ctx, &dynamodb.ScanInput{TableName: aws.String(deleteRecordTableName)})
-	require.NoError(t, err)
-	// All delete records have been removed
-	assert.Zero(t, scanOut.ScannedCount)
-	assert.Empty(t, scanOut.Items)
 
 }
 
@@ -328,7 +360,7 @@ func (s *TestSourcePackage) WithSources(count int, bucketName string, fileConfig
 	return s
 }
 
-func (s *TestSourcePackage) Insert(ctx context.Context, db store.TestDB, orgId int) *TestSourcePackage {
+func (s *TestSourcePackage) Insert(ctx context.Context, db *store.TestDB, orgId int) *TestSourcePackage {
 	s.Package.Insert(ctx, db, orgId)
 	for _, f := range s.Files {
 		f.Insert(ctx, db, orgId)
@@ -336,26 +368,32 @@ func (s *TestSourcePackage) Insert(ctx context.Context, db store.TestDB, orgId i
 	return s
 }
 
+// PutObjectInputs returns a slice of *s3.PutObjectInput for each unpublished file in this test package.
 func (s *TestSourcePackage) PutObjectInputs() []*s3.PutObjectInput {
 	var inputs []*s3.PutObjectInput
-	for i, f := range s.Files {
-		inputs = append(inputs, &s3.PutObjectInput{
-			Bucket: aws.String(f.S3Bucket),
-			Key:    aws.String(f.S3Key),
-			Body:   strings.NewReader(fmt.Sprintf("object %d content", i)),
-		})
+	for _, f := range s.Files {
+		if !f.Published {
+			inputs = append(inputs, &s3.PutObjectInput{
+				Bucket: aws.String(f.S3Bucket),
+				Key:    aws.String(f.S3Key),
+				Body:   strings.NewReader(strings.Repeat("a", int(f.Size))),
+			})
+		}
 	}
 	return inputs
 }
 
+// DeleteFiles places a delete marker on each unpublished file in this test package. It returns a map from s3 key to the S3 version id of the delete marker.
 func (s *TestSourcePackage) DeleteFiles(ctx context.Context, t require.TestingT, s3Client *s3.Client) map[string]string {
 	s3KeyToDeleteMarkerVersion := map[string]string{}
 	for _, f := range s.Files {
-		// Delete the object from S3
-		deleteObjectOutput, err := s3Client.DeleteObject(ctx, &s3.DeleteObjectInput{Bucket: aws.String(f.S3Bucket), Key: aws.String(f.S3Key)})
-		require.NoError(t, err, "error setting up deleted objects")
-		require.NotNil(t, deleteObjectOutput.VersionId, "version id of delete is nil; is bucket versioning enabled?")
-		s3KeyToDeleteMarkerVersion[f.S3Key] = aws.ToString(deleteObjectOutput.VersionId)
+		if !f.Published {
+			// Delete the object from S3
+			deleteObjectOutput, err := s3Client.DeleteObject(ctx, &s3.DeleteObjectInput{Bucket: aws.String(f.S3Bucket), Key: aws.String(f.S3Key)})
+			require.NoError(t, err, "error setting up deleted objects")
+			require.NotNil(t, deleteObjectOutput.VersionId, "version id of delete is nil; is bucket versioning enabled?")
+			s3KeyToDeleteMarkerVersion[f.S3Key] = aws.ToString(deleteObjectOutput.VersionId)
+		}
 	}
 	return s3KeyToDeleteMarkerVersion
 }
@@ -372,4 +410,11 @@ func (s *TestSourcePackage) PutItemInputs(t require.TestingT, deleteRecordTableN
 		})
 	}
 	return putItemInputs
+}
+
+func (s *TestSourcePackage) Size() (size int64) {
+	for _, f := range s.Files {
+		size += f.Size
+	}
+	return
 }
